requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
openpyxl==3.1.2
sqlalchemy==2.0.23
python-multipart==0.0.6
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
pydantic==2.5.2
email-validator==2.1.1
streamlit-oauth==0.1.14
psycopg2-binary==2.9.9